chore: extract navigation tenant selection content
