+++ src/components/navigation/header/navigationTenantMenu/Loading.tsx
@@ -7,9 +7,10 @@ export const NavigationTenantMenuLoading: FC = () => (
   <>
     <Box
       position="absolute"
-      w="full"
-      h="full"
       top="0"
+      bottom="0"
+      left="0"
+      right="0"
       bg="gray.900"
       opacity="0.7"
     />
